undo rename to keep changes easier to review
