+++ backend/molgenis-emx2-beacon-v2/src/main/java/org/molgenis/emx2/beaconv2/common/AgeAndAgeGroup.java
@@ -1,7 +1,7 @@
 package org.molgenis.emx2.beaconv2.common;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import org.molgenis.emx2.rdf.OntologyTerm;
+import org.molgenis.emx2.semantics.OntologyTerm;
 
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
 public class AgeAndAgeGroup {
