+++ backend/molgenis-emx2-beacon-v2/src/main/java/org/molgenis/emx2/beaconv2/endpoints/runs/RunsResponse.java
@@ -1,6 +1,6 @@
 package org.molgenis.emx2.beaconv2.endpoints.runs;
 
-import static org.molgenis.emx2.rdf.QueryHelper.mapToOntologyTerm;
+import static org.molgenis.emx2.semantics.QueryHelper.mapToOntologyTerm;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonIgnore;
