+++ backend/molgenis-emx2-fairdatapoint/src/main/java/org/molgenis/emx2/fairdatapoint/FAIRDataPointCatalog.java
@@ -2,9 +2,9 @@ package org.molgenis.emx2.fairdatapoint;
 
 import static org.eclipse.rdf4j.model.util.Values.iri;
 import static org.eclipse.rdf4j.model.util.Values.literal;
-import static org.molgenis.emx2.rdf.IRIParsingEncoding.encodedIRI;
-import static org.molgenis.emx2.rdf.IRIParsingEncoding.getURI;
-import static org.molgenis.emx2.rdf.RDFService.extractHost;
+import static org.molgenis.emx2.semantics.IRIParsingEncoding.encodedIRI;
+import static org.molgenis.emx2.semantics.IRIParsingEncoding.getURI;
+import static org.molgenis.emx2.semantics.RDFService.extractHost;
 
 import graphql.ExecutionResult;
 import graphql.GraphQL;
