+++ backend/molgenis-emx2-fairdatapoint/src/main/java/org/molgenis/emx2/fairdatapoint/FAIRDataPointDataset.java
@@ -5,9 +5,9 @@ import static org.eclipse.rdf4j.model.util.Values.literal;
 import static org.molgenis.emx2.fairdatapoint.FAIRDataPointCatalog.extractItemAsIRI;
 import static org.molgenis.emx2.fairdatapoint.FormatMimeTypes.FORMATS;
 import static org.molgenis.emx2.fairdatapoint.Queries.queryDataset;
-import static org.molgenis.emx2.rdf.IRIParsingEncoding.encodedIRI;
-import static org.molgenis.emx2.rdf.IRIParsingEncoding.getURI;
-import static org.molgenis.emx2.rdf.RDFService.extractHost;
+import static org.molgenis.emx2.semantics.IRIParsingEncoding.encodedIRI;
+import static org.molgenis.emx2.semantics.IRIParsingEncoding.getURI;
+import static org.molgenis.emx2.semantics.RDFService.extractHost;
 
 import java.io.StringWriter;
 import java.net.URI;
