+++ backend/molgenis-emx2-semantics/src/main/java/org/molgenis/emx2/semantics/OntologyTerm.java
@@ -1,4 +1,4 @@
-package org.molgenis.emx2.rdf;
+package org.molgenis.emx2.semantics;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonIgnore;
