+++ backend/molgenis-emx2-semantics/src/test/java/org/molgenis/emx2/semantics/OntologyTableSemanticsTest.java
@@ -1,4 +1,4 @@
-package org.molgenis.emx2.rdf;
+package org.molgenis.emx2.semantics;
 
 import static org.junit.jupiter.api.Assertions.*;
