+++ settings.gradle
@@ -63,7 +63,7 @@ include(':backend:molgenis-emx2-webapi')
 include(':backend:molgenis-emx2-graphql')
 include(':backend:molgenis-emx2-run')
 include(':backend:molgenis-emx2-tasks')
-include(':backend:molgenis-emx2-rdf')
+include(':backend:molgenis-emx2-semantics')
 include(':backend:molgenis-emx2-beacon-v2')
 include(':backend:molgenis-emx2-fairdatapoint')
 include(':backend:molgenis-emx2-graphgenome')
